Fix the simple example for ARM architecture integers

+++ examples/simple.c
@@ -5,28 +5,28 @@
  *
  * Written after the first test was defined:
  *
- * long add(int a, int b)
+ * int add(int a, int b)
  * {
  *   return 0;
  * }
  *
  * Written after the second test was defined:
  *
- * long add(int a, int b)
+ * long long add(int a, int b)
  * {
  *   return a + b;
  * }
  *
  * Written after the third and fourth test was defined:
  *
- * long add(int a, int b)
+ * long long add(int a, int b)
  * {
- *   return (long)a + (long)b;
+ *   return (long long)a + (long)b;
  * }
  *
  */
 
-long add(int a, int b)
+long long add(int a, int b)
 {
-  return (long)a + (long)b;
+  return (long long)a + (long long)b;
 }
